# Make byte tensor serialization faster for python client