This pr improves byte tensor serialization by order of magnitude. (`6.5s` -> `10.4 ms`), thus making `set_data_from_numpy` for byte tensors much faster. 

### Todo:
- [x] Ensure tests and build pass

### Perf Comparision
```python
import tritonclient.grpc as grpcclient
import numpy as np


log_ls = ['test log '*100]*10000
log_ls = [l.encode('utf-8') for l in log_ls]
log_ar = np.array(log_ls).reshape(1,len(log_ls))

input_grpc = grpcclient.InferInput("raw_log",log_ar.shape,"BYTES")
```

##### Without this PR:
```python
input_grpc.set_data_from_numpy(log_ar)
```
```
Wall time:  6.55 s
```

##### With this PR:
```python
input_grpc.set_data_from_numpy(log_ar)
```
```
CPU times: user 9.24 ms, sys: 1.26 ms, total: 10.5 ms
Wall time: 10.4 ms
```

CC: @BartleyR, @bsuryadevara