@VibhuJawa 2021-03-31
@tanmayv25 , 

If this change is acceptable we should also target it to the `python_backend`. 

https://github.com/triton-inference-server/python_backend/blob/15ca4cf1d3336844d88ab3f31b930639d47dfd4d/src/resources/startup.py#L54-L108