@Tabrizian 2021-03-31
Good catch and thank you for the PR. What do you think is the underlying reason for this speedup? Is it because `bytes` is immutable and lists are mutable? Do you think using `bytearray` will lead to the same speedup?