@VibhuJawa 2021-03-31
> Good catch and thank you for the PR. What do you think is the underlying reason for this speedup? Is it because `bytes` is immutable and lists are mutable? Do you think using `bytearray` will lead to the same speedup?

@Tabrizian,

Kind of, I think this is because of the number of memory allocations that are happening as well as the increase in size of allocations. 

If we have something like below , we have 2*n big allocation and 2*n small allocation and the cost of each subsequence allocation keeps on increasing. 

```python
flattened = bytes()
s = 'test string' * 1024
n = 1000

for _ in range(n):
   flattened =  flattened + struct.pack("<I", len(s)) ## 2 allocations , 1 big , 1 small 
   flattened = flattened + s ## 2 allocations , 1 big , 1 small 
``` 
```
1.19 s ± 26.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
``` 

### Graph of cost of  allocations (increases incrementally due to increase increase in size of allocation) 
```
flattened = flattened + s 
```

![image](https://user-images.githubusercontent.com/4837571/113189726-5d74cd80-9210-11eb-982e-4da74412c126.png)


          
### Alternate Solution: 
This should just have just 2n small allocations and 1 big allocation  :
```python

s = 'test string' * 1024
n = 1000
flattened_ls = []

for _ in range(n):
  flattened_ls.append(struct.pack("<I", len(s))) # 1 small allocation 
  flattened_ls.append(s) ## 1 small allocation 
  
flattened = b''.join(flattened_ls) ## 1 big allocation 
```
```
1.27 ms ± 43.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```